fixing!

+++ rrt_planner_point_robot.py
@@ -25,35 +25,29 @@ import math
 
 def redraw(canvas):
     canvas.clear()
-    canvas.markit(tx, ty, r=SMALLSTEP)
+    canvas.markit( tx, ty, r=SMALLSTEP )
     drawGraph(G, canvas)
-    for o in obstacles:
-        canvas.showRect(o, outline='blue', fill='blue')
+    for o in obstacles: canvas.showRect(o, outline='blue', fill='blue')
     canvas.delete("debug")
 
-
 def drawGraph(G, canvas):
-    global vertices, nodes, edges
-    if not visualize:
-        return
+    global vertices,nodes,edges
+    if not visualize: return
     for i in G[edges]:
-        # e.g. vertices: [[10, 270], [10, 280]]
-        canvas.polyline([vertices[i[0]], vertices[i[1]]])
+       canvas.polyline(  [vertices[i[0]], vertices[i[1]] ]  )
 
-
-# Use this function to generate points randomly for the RRT algo
 def genPoint():
-    # if args.rrt_sampling_policy == "uniform":
-    #     # Uniform distribution
-    #     x = random.random()*XMAX
-    #     y = random.random()*YMAX
-    # elif args.rrt_sampling_policy == "gaussian":
-    #     # Gaussian with mean at the goal
-    #     x = random.gauss(tx, sigmax_for_randgen)
-    #     y = random.gauss(ty, sigmay_for_randgen)
-    # else:
-    #     print ("Not yet implemented")
-    #     quit(1)
+    if args.rrt_sampling_policy == "uniform":
+        # Uniform distribution
+        x = random.random()*XMAX
+        y = random.random()*YMAX
+    elif args.rrt_sampling_policy == "gaussian":
+        # Gaussian with mean at the goal
+        x = random.gauss(tx, sigmax_for_randgen)
+        y = random.gauss(ty, sigmay_for_randgen)
+    else:
+        print ("Not yet implemented")
+        quit(1)
 
     bad = 1
     while bad:
@@ -67,142 +61,143 @@ def genPoint():
             x = random.gauss(tx, sigmax_for_randgen)
             y = random.gauss(ty, sigmay_for_randgen)
         else:
-            print("Not yet implemented")
+            print ("Not yet implemented")
             quit(1)
         # range check for gaussian
-        if x < 0:
-            bad = 1
-        if y < 0:
-            bad = 1
-        if x > XMAX:
-            bad = 1
-        if y > YMAX:
-            bad = 1
-    return [x, y]
-
+        if x<0: bad = 1
+        if y<0: bad = 1
+        if x>XMAX: bad = 1
+        if y>YMAX: bad = 1
+    return [x,y]
 
 def returnParent(k, canvas):
     """ Return parent note for input node k. """
     for e in G[edges]:
-        if e[1] == k:
-            canvas.polyline([vertices[e[0]], vertices[e[1]]], style=3)
+        if e[1]==k:
+            canvas.polyline(  [vertices[e[0]], vertices[e[1]] ], style=3  )
             return e[0]
-            
 
 def genvertex():
-    vertices.append(genPoint())
+    vertices.append( genPoint() )
     return len(vertices)-1
-
 
 def pointToVertex(p):
-    vertices.append(p)
+    vertices.append( p )
     return len(vertices)-1
 
-
 def pickvertex():
-    return random.choice(range(len(vertices)))
+    return random.choice( range(len(vertices) ))
 
-
-def lineFromPoints(p1, p2):
-    return (p1[1]-p2[1])/(p1[0]-p2[0])
-
+def lineFromPoints(p1,p2):
+    line = []
+    llsq = 0.0 # line length squared
+    for i in range(len(p1)):  # each dimension
+        h = p2[i] - p1[i]
+        line.append( h )
+        llsq += h*h
+    ll = math.sqrt(llsq)  # length
+    # normalize line
+    if ll <=0: return [0,0]
+    for i in range(len(p1)):  # each dimension
+        line[i] = line[i]/ll
+    return line
 
 def pointPointDistance(p1, p2):
     dist = ((p2[1]-p1[1])**2+(p2[0]-p1[0])**2)**0.5
     return dist
 
-
-def closestPointToPoint(G, p2):
-    nNear = 0
-    minDist = pointPointDistance(vertices[0], p2)
-    for node in G[nodes]:
-        dist = pointPointDistance(vertices[node], p2)
-        if dist < minDist:
-            dist = minDist
-            nNear = node
-    return nNear
-
-
-#return true if slope of p3p1 > p2p1
-def checkSlope(p1,p2,p3):
-    return (p3[1]-p1[1]) / (p3[0]-p1[0]) > (p2[1]-p1[1]) / (p2[0]-p1[0])
-
-# Return true if line segments p1p2 and p3p4 intersect
-def intersect(p1,p2,p3,p4):
-    return checkSlope(p1,p2,p4) != checkSlope(p2,p3,p4) and checkSlope(p1,p2,p3) != checkSlope(p1,p2,p4)
-    
-def lineHitsRect(p1, p2, r):
-    i1 = intersect(p1,p2,(r[0],r[3]),(r[0],r[1]))
-    i2 = intersect(p1,p2,(r[2],r[3]),(r[2],r[1]))
-    i3 = intersect(p1,p2,(r[0],r[1]),(r[2],r[1]))
-    i4 = intersect(p1,p2,(r[2],r[3]),(r[0],r[3]))
-    if(i1 or i2 or i3 or i4):
-        return 1
-    return 0
+def closestPointToPoint(G,p2):
+    dmin = 999999999
+    for v in G[nodes]:
+        p1 = vertices [ v ]
+        d = pointPointDistance(p1,p2)
+        if d <= dmin:
+            dmin = d
+            close = v
+    return close
 
 
-def inRect(p, rect, dilation):
-    if p[0] < rect[0]-dilation:
-        return 0
-    if p[1] < rect[1]-dilation:
-        return 0
-    if p[0] > rect[2]+dilation:
-        return 0
-    if p[1] > rect[3]+dilation:
-        return 0
-    return 1
+#used in intersect
+def ccw(A,B,C):
+    """ Determine if three points are listed in a counterclockwise order.
+    For three points A, B and C. If the slope of the line AB is less than
+    the slope of the line AC then the three points are in counterclockwise order.
+    See:  http://compgeom.cs.uiuc.edu/~jeffe/teaching/373/notes/x06-sweepline.pdf
+    """
+    return (C[1]-A[1])*(B[0]-A[0]) > (B[1]-A[1])*(C[0]-A[0])
+
+#used in lineHitsRect
+def intersect(A,B,C,D):
+        """ do lines AB and CD intersect? """
+        i = ccw(A,C,D) != ccw(B,C,D) and ccw(A,B,C) != ccw(A,B,D)
+        #if i:
+        #    canvas.polyline(  [ A,B ], style=4  , tags = ("debug"))
+        #    canvas.polyline(  [ C,D ], style=4  , tags = ("debug"))
+        #else:
+        #    canvas.polyline(  [ A,B ], style=1  , tags = ("debug")) # green
+        #    canvas.polyline(  [ C,D ], style=1  , tags = ("debug"))
+        return i
 
 
-def addNewPoint(p1, p2, stepsize):
-#    theta = math.atan(lineFromPoints(p1,p2))
-    dist = pointPointDistance(p1,p2)
-#   *math.cos(theta), *math.sin(theta)
-    x = p1[0] + stepsize*(p2[0]-p1[0])/dist
-    y = p1[1] + stepsize*(p2[1]-p1[1])/dist
-    return (x,y)
+def lineHitsRect(p1,p2,r):
+   rline = ( (r[0],r[1]), (r[0],r[3]) )
+   if intersect( p1, p2, rline[0], rline[1] ): return 1
+   rline = ( (r[0],r[1]), (r[2],r[1]) )
+   if intersect( p1, p2, rline[0], rline[1] ): return 1
+   rline = ( (r[0],r[3]), (r[2],r[3]) )
+   if intersect( p1, p2, rline[0], rline[1] ): return 1
+   rline = ( (r[2],r[1]), (r[2],r[3]) )
+   if intersect( p1, p2, rline[0], rline[1] ): return 1
 
-def hitsObstacles(obs, newp, p):
-    for o in obs:
-        # The following function defined by you must handle the occlusion cases
-        if lineHitsRect(newp, p, o) or inRect(newp, o, 1):
-            return False
-    return False
+   return 0
 
+def inRect(p,rect,dilation):
+   """ Return 1 if p is inside rect, dilated by dilation (for edge cases). """
+   if p[0]<rect[0]-dilation: return 0
+   if p[1]<rect[1]-dilation: return 0
+   if p[0]>rect[2]+dilation: return 0
+   if p[1]>rect[3]+dilation: return 0
+   return 1
 
 def rrt_search(G, tx, ty, canvas):
     global sigmax_for_randgen, sigmay_for_randgen
-    n = 0
-    nsteps = 0
-    while 1:  # Main loop
-        # This generates a point in form of [x,y] from either the normal dist or the Gaussian dist
-        p = genPoint()
-
-        # This function must be defined by you to find the closest point in the existing graph to the guiding point
-        k = closestPointToPoint(G, p)
-        cp = vertices[k]
-        newPoint = addNewPoint(cp, p, SMALLSTEP)
+    n=0
+    nsteps=0
+    while 1:
+        x_random = genPoint()
+        if n%100 == 0: x_random = (tx,ty)
+        v = closestPointToPoint(G,x_random)
+        x_nearest = vertices[v]
+        line_between = lineFromPoints(x_nearest,x_random)
+        move_distance = [line_between[0] * SMALLSTEP, line_between[1] * SMALLSTEP]
+        x_new = [x_nearest[0]+move_distance[0], x_nearest[1]+move_distance[1]]
 
         if visualize:
-#            if nsteps%500 == 0: redraw(canvas)  # erase generated points now and then or it gets too cluttered
-            n = n+1
-            if n > 10:
+            # if nsteps%500 == 0: redraw()  # erase generated points now and then or it gets too cluttered
+            n=n+1
+            if n>10:
                 canvas.events()
-                n = 0
+                n=0
 
-        if not hitsObstacles(obstacles,newPoint, cp):
-            v = pointToVertex(newPoint)
-            G[nodes].append(v)
-            G[edges].append((k,v))
+        in_rect = 0
+        for o in obstacles :
+            if inRect(x_new,o,1) or lineHitsRect(x_nearest,x_new,o):
+                in_rect = 1
+
+        if not in_rect:
+            k = pointToVertex( x_new )   # is the new vertex ID
+            G[nodes].append(k)
+            G[edges].append( (v,k) )
             if visualize:
-                canvas.polyline([vertices[v], vertices[k]])
-#                canvas.markit(p[0], p[1], r=1)
-            if pointPointDistance(newPoint, [tx, ty]) < SMALLSTEP:
-                print("Target achieved.", nsteps, "nodes in entire tree")
+                canvas.polyline(  [x_nearest, x_new] )
+
+            if pointPointDistance( x_new, [tx,ty] ) < SMALLSTEP:
+                print ("Target achieved.", nsteps, "nodes in entire tree")
                 if visualize:
                     t = pointToVertex([tx, ty])  # is the new vertex ID
                     G[edges].append((k, t))
                     if visualize:
-                        canvas.polyline([p, vertices[t]], 1)
+                        canvas.polyline([x_new, vertices[t]], 1)
                     # while 1:
                     #     # backtrace and show the solution ...
                     #     canvas.events()
@@ -212,12 +207,10 @@ def rrt_search(G, tx, ty, canvas):
                         oldp = vertices[k]  # remember point to compute distance
                         k = returnParent(k, canvas)  # follow links back to root.
                         canvas.events()
-                        if k <= 1:
-                            break  # have we arrived?
+                        if k <= 1: break  # have we arrived?
                         nsteps = nsteps + 1  # count steps
-                        totaldist = totaldist + \
-                            pointPointDistance(vertices[k], oldp)  # sum lengths
-                    print("Path length", totaldist, "using", nsteps, "nodes.")
+                        totaldist = totaldist + pointPointDistance(vertices[k], oldp)  # sum lengths
+                    print ("Path length", totaldist, "using", nsteps, "nodes.")
 
                     global prompt_before_next
                     if prompt_before_next:
@@ -225,15 +218,10 @@ def rrt_search(G, tx, ty, canvas):
                         print("More [c,q,g,Y]>")
                         d = sys.stdin.readline().strip().lstrip()
                         print("[" + d + "]")
-                        if d == "c":
-                            canvas.delete()
-                        if d == "q":
-                            return
-                        if d == "g":
-                            prompt_before_next = 0
+                        if d == "c": canvas.delete()
+                        if d == "q": return
+                        if d == "g": prompt_before_next = 0
                     break
-            nsteps += 1
-
 
 def main():
     # seed
@@ -253,8 +241,8 @@ def main():
         drawGraph(G, canvas)
         rrt_search(G, tx, ty, canvas)
 
-        if visualize:
-            canvas.mainloop()
+    if visualize:
+        canvas.mainloop()
 
 
 if __name__ == '__main__':
